pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-web-skill-adapter"
version = "0.1.0"
description = "ADK demo that discovers remote web skills and exposes them as dynamic tools."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-adk",
  "httpx>=0.27,<1",
]

[project.scripts]
web-skill-demo = "web_skill_adapter.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["web_skill_adapter*"]
